d.2.1-1

+++ src/app/pages/time/time.page.ts
@@ -5,6 +5,7 @@ import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { OfficeService } from 'src/app/services/office/office.service';
 import { SeoService } from 'src/app/services/seo/seo.service';
 import { CustomDate } from 'src/app/models/class/date/custom-date';
+import { AlertService } from 'src/app/services/alert/alert.service';
 
 @Component({
   selector: 'app-time',
@@ -36,7 +37,8 @@ export class TimePage implements OnInit {
   constructor(
     private fb: FormBuilder,
     private seoService: SeoService,
-    private officeService: OfficeService
+    private officeService: OfficeService,
+    private alertService :AlertService
   ) { }
   workForm: FormGroup = this.fb.group({
     createdAt: [serverTimestamp()],
@@ -73,8 +75,11 @@ export class TimePage implements OnInit {
     this.officeService.addWork(this.workForm.value)
   }
 
-  deleteWork(idField: string) {
-    this.officeService.deleteWork(idField)
+  async deleteWork(idField: string) {
+    const response = await this.alertService.deleteAlert()
+    if ( response == "confirm") {
+      this.officeService.deleteWork(idField)
+    }
   }
 
   async copyToClipboard(work: any) {
